reorganized code structures
