+++ deploy/python/resources/configuration/example_dodoor_configuration.conf
